+++ src/main.ts
@@ -1,7 +1,9 @@
 import { Plugin } from "obsidian";
 import { registerLibraryCommands } from "./commands/libraryCommands";
 import { KnowledgeLibraryPluginSettings } from "./core/settings";
+import { ResourceService } from "./services/ResourceService";
 import { TagAliasService } from "./services/TagAliasService";
+import { TagService } from "./services/TagService";
 import { RibbonService } from "./ui/RibbonService";
 import { StatusBarService } from "./ui/StatusBarService";
 
@@ -12,6 +14,8 @@ const DEFAULT_SETTINGS: KnowledgeLibraryPluginSettings = {
 export default class KnowledgeLibraryPlugin extends Plugin {
   settings: KnowledgeLibraryPluginSettings = DEFAULT_SETTINGS;
   tagAliases!: TagAliasService;
+  tagService!: TagService;
+  resourceService!: ResourceService;
   private ribbonService!: RibbonService;
   private statusBarService!: StatusBarService;
 
@@ -19,6 +23,8 @@ export default class KnowledgeLibraryPlugin extends Plugin {
     await this.loadSettings();
 
     this.tagAliases = new TagAliasService();
+    this.tagService = new TagService(this.tagAliases);
+    this.resourceService = new ResourceService(undefined, this.tagService);
     this.ribbonService = new RibbonService(this);
     this.statusBarService = new StatusBarService(this, this.settings.versionLabel);